chore: REMAKE 재생성,마지막씬 수정 완료

+++ server/Amuse/src/main/java/com/muse/amuze/novel/model/dto/ChatMessageRequest.java
@@ -14,5 +14,5 @@ public class ChatMessageRequest {
   private String roomMode;
   private String autoMode;
   private String userInput;
-  private Integer lastSceneId;
+  private Long lastSceneId;
 }
